feat: Make target cards expandable like history cards.
