+++ pkg/webui/ui/src/components/targets-view/Card.tsx
@@ -1,5 +1,8 @@
 import React from "react";
-import { Box, BoxProps, Paper, PaperProps, Tooltip, Typography } from "@mui/material"
+import { Box, BoxProps, Divider, IconButton, Paper, PaperProps, Tab, Tooltip, Typography } from "@mui/material"
+import { CloseLightIcon } from "../../icons/Icons";
+import { SidePanelProvider, useSidePanelTabs } from "../result-view/SidePanel";
+import { TabContext, TabList, TabPanel } from "@mui/lab";
 
 export const cardWidth = 247;
 export const projectCardMinHeight = 80;
@@ -11,12 +14,13 @@ export const CardPaper = React.forwardRef((props: PaperProps, ref: React.Forward
     return <Paper
         elevation={5}
         sx={{
-            width: cardWidth,
-            height: cardHeight,
+            width: '100%',
+            height: '100%',
             borderRadius: '12px',
             border: '1px solid #59A588',
             boxShadow: '4px 4px 10px #1E617A',
             flexShrink: 0,
+            position: 'relative',
             ...sx
         }}
         {...rest}
@@ -44,7 +48,9 @@ export const CardTemplate = React.forwardRef((props: {
     subheader?: string,
     subheaderTooltip?: React.ReactNode,
     body?: React.ReactNode,
-    footer?: React.ReactNode
+    footer?: React.ReactNode,
+    showCloseButton?: boolean,
+    onClose?: () => void
 }, ref: React.ForwardedRef<HTMLDivElement>) => {
     const icon = props.icon && (
         <Tooltip title={props.iconTooltip}>
@@ -85,6 +91,7 @@ export const CardTemplate = React.forwardRef((props: {
                 fontSize='14px'
                 fontWeight={500}
                 lineHeight='19px'
+                width='max-content'
             >
                 {props.subheader}
             </Typography>
@@ -92,7 +99,7 @@ export const CardTemplate = React.forwardRef((props: {
     );
 
     const body = props.body && (
-        <Box flex='1 1 auto'>{props.body}</Box>
+        <Box flex='1 1 auto' overflow='hidden' padding='0 16px'>{props.body}</Box>
     );
 
     const footer = props.footer && (
@@ -102,11 +109,23 @@ export const CardTemplate = React.forwardRef((props: {
     );
 
     return <CardPaper {...props.paperProps} ref={ref}>
+        {props.showCloseButton && (
+            <Box
+                position='absolute'
+                right='10px'
+                top='10px'
+            >
+                <IconButton onClick={props.onClose}>
+                    <CloseLightIcon />
+                </IconButton>
+            </Box>
+        )}
         <Box
             display='flex'
             flexDirection='column'
             justifyContent='space-between'
             height='100%'
+            gap='10px'
             {...props.boxProps}
         >
             <Box flex='0 0 auto' display='flex' gap='15px'>
@@ -123,3 +142,36 @@ export const CardTemplate = React.forwardRef((props: {
 });
 
 CardTemplate.displayName = 'CardTemplate';
+
+export const CardBody = React.memo((props: { provider: SidePanelProvider }) => {
+    const { tabs, selectedTab, handleTabChange } = useSidePanelTabs(props.provider)
+
+    if (!props.provider
+        || !selectedTab
+        || !tabs.find(x => x.label === selectedTab)
+    ) {
+        return null;
+    }
+
+    return <TabContext value={selectedTab}>
+        <Box display='flex' flexDirection='column' height='100%' overflow='hidden'>
+            <Box height='36px' flex='0 0 auto' p='0'>
+                <TabList onChange={handleTabChange}>
+                    {tabs.map((tab, i) => {
+                        return <Tab label={tab.label} value={tab.label} key={tab.label} />
+                    })}
+                </TabList>
+            </Box>
+            <Divider sx={{ margin: 0 }} />
+            <Box overflow='auto' p='10px 0'>
+                {tabs.map(tab => {
+                    return <TabPanel key={tab.label} value={tab.label} sx={{ padding: 0 }}>
+                        {tab.content}
+                    </TabPanel>
+                })}
+            </Box>
+        </Box>
+    </TabContext>
+});
+
+CardBody.displayName = 'CardBody';
