+++ pkg/webui/ui/src/components/targets-view/ProjectItem.tsx
@@ -3,7 +3,7 @@ import { Box } from "@mui/material";
 import React from "react";
 import { ProjectIcon } from "../../icons/Icons";
 import { ProjectSummary } from "../../project-summaries";
-import { CardTemplate, projectCardMinHeight } from "./Card";
+import { CardTemplate, cardWidth, projectCardMinHeight } from "./Card";
 
 export const ProjectItem = React.memo((props: { ps: ProjectSummary }) => {
     const name = getLastPathElement(props.ps.project.gitRepoKey)
@@ -20,6 +20,7 @@ export const ProjectItem = React.memo((props: { ps: ProjectSummary }) => {
         paperProps={{
             sx: {
                 padding: '20px 16px',
+                width: cardWidth,
                 height: 'auto',
                 minHeight: projectCardMinHeight
             }
